chore : lib> checklist 폴더 분리
